This commit fixes a bug where errors are not visible to end users.
A billing simulation is designed to help a user identify errors before submitting to the Ministry of Health.

In a production database, a non-doctor provider in the database had a test billing code of 12345, which is not the right number of characters. When running a simulation, the user was told “error count 1”, but there wasn’t any information provided as to what that error was, and so it wasn’t possible to address the issue.

Two issues have been addressed in this pull request:

a) The errors for providers that have no invoices (therefore dbObj.getRecordCount() == 0) end up in the error count, but they are not visible to the user on the GUI. The user therefore is told there is an error, but they don’t have any idea what the error is.

b) Some of the error messages are not very descriptive. “The Provider’s billing code is not correct”…. doesn’t tell you which provider has a billing code that isn’t correct.