For this jsp, there was recently a jquery ui update from 1.8 to 1.12
This change has created regression where for a saved consult request, when it is opened again, then the "add attachments" clicked, the currently selected attachments are not pre-selected.

The reason is because there were changes to the complete callback behavior in jQuery UI between versions 1.8 and 1.12. Specifically, the complete callback was removed from the show option in jQuery UI 1.9. In jQuery UI 1.8 and earlier versions, the show option supported a complete callback that would be triggered after the dialog finished showing. However, starting from jQuery UI 1.9, the complete callback was no longer available in the show option. If you need to run code after the dialog finishes showing in jQuery UI 1.9 and later versions, you can use the open event instead:

The code was slightly refactored so that only after the list of documents has been successfully loaded would the code that was formerly in the "complete" function executed, this code is what "pre-selecting" the attachments.  The timing of the slide open animation was also slightly tweaked so that the data would have a bit more time to load successfully, before the dialog fully opened.  1 second is believed to be enough time to load the list of documents/labs in most production instances.