fixing bug where if there's a row in waitingListName (and so a waiting list is returned), but the property DEMOGRAPHIC_WAITING_LIST is set to false, attempting to update a patient's demographic results in a 404 error.
the fundamental issue is very simple - in demographiceditdemographic.jsp, the property DEMOGRAPHIC_WAITING_LIST is explicitly checked and various fields are hidden if that property is false.  however, on the actual update, the property is not getting checked, instead, whether there's a waiting list was the determining factor.

in OSCAR instances where there's a waiting list (E.g. for historical reasons, for test reasons), but now where the property is set to false, attempting to update a demographic will result in an error.    Adding an explicit check of the property addresses this issue.