aider: Created interface and implementation classes for SecRoleDao.
# Aider chat conversation:

USER: I am updating my code from Hibernate 3 to Hibernate 4. To do this, we need to implement interfaces for DAO (Data Access Object) classes. I have already implemented this in some files, like DemographicExtArchiveDao for DemographicExtArchiveDaoImpl. Now, I need you to create interfaces and implementation classes for the remaining DAO classes that currently lack both an interface and an implementation class. Here are the guidelines for making these changes: 1. Ensure that the name of the implementation class ends with 'Impl' and the interface name ends with 'Dao'. For example, an interface class should be named ContactDao.java and its implementation should be ContactDaoImpl.java. 2. All files should include the package 'org.oscarehr.common.dao' at the top of the file. 3. If an implementation class extends the AbstractDaoImpl class, then the corresponding interface should extend the AbstractDao interface. 4. Apply these changes to all DAO classes in the java files 5. Do this for SecRoleDao.java

ASSISTANT: No changes are needed.