feat(fix): Removed static from the getCustomPreventionItems() method to allow for it to be called properly.
This fix ultimately allows the "Prevention List Manager" page to be loaded from the System Management dropdown
fixing its implementation in the administration panel.

In all the places the method is being called, it is not being called directly from the classname allowing for
this change to be implemented properly without issue.